Jack AUX Ports Guaranteed to Correspond to Program AUX Outputs
The correspondence is only guaranteed on the Jack side.  While
it is not handled explicitly at the moment (but probably should be)
if an program AUX port fails to create a jack port, then it should
simply not be listed in the thread shared map.

AUX port names use the tile handle, which is monotonic and unique.

AUX port names are not stable between patch saves and loads, but
their ordering is expected to be.