Changed display selection to be manual, added DPI override mechanism.
sys.argv[1] now can specify the display index if you need to put
mollytime on a specific monitor.

sys.argv[2] now can override the vertical size of the screen in inches,
which circumvents automatic DPI calculation.

The former avoids the problem of Gnome 48 / Wayland breaking fullscreen
display selection when the primary display has a fractional resolution
like 175%.

The latter provides a mediocre fallback for when the system reports a
bad value or the sad day when the Wayland people decide to break xrandr.