Pull Alsa MIDI implementation into an abstract `MidiDriver`
The MIDI interface now passes-through to a `MidiDriver` implementation, selected via conditional compilation. Alsa is now provided as the (only) implementation.

The driver is a uniquely owned singleton, in line with the program's behavior elsewhere. The only implementation requirement is `ProcessEvents()`; any MidiDriver initialization & de-initialization should be handled idiomatically in constructors / destructors.

It's OK to not have a MIDI driver. If nothing is found, the system will make note, and proceed gracefully.