Refactor audio backend to isolate Jack implementation
AudioStream is now a pure-virtual interface for a specific backend to implement. The top-level interface now lives in an `Audio` namespace, which (like the new MIDI architecture) holds a uniquely-owned AudioStream implementation, selected by conditional compilation.

The Jack implementation required some rearchitecture to properly encapsulate into an independent object. Setup and teardown logic has been moved into constructors & destructors. (This requires a little initializer-list maneuvering, since the JackRealTimeThread has to be constructed after the JackClient is initialized.) Otherwise, this is just moving all the state into one place, and trimming a lot of now-unnecessary checks and re-checks on singleton initialization.

Note: I've removed the `jack_on_shutdown()` call, as it doesn't appear to be necessary. The client will close whenever the JackStream's lifetime expires (whether naturally or via exception), and the documentation notes that client's "aren't expected" to need this anyway.