Clean up JackStream initialization
JACK Initialization machinery was previously split between the JackStream implementation, and its JackRealTimeThread member, but these initialization steps have:
1. A strong order dependency...
2. For data that's logically owned by the holder of the JACK client and its `BufferState` (i.e. JackStream)...
3. On which the JackRealTimeThread doesn't actually directly depend, as it instead goes exclusively through a non-owning reference to the `BufferState`.

Accounting for this as-is required some tricky machinery with the JackStream initializer list. Instead, it's much simpler to have the JackStream constructor handle all this at once, and limit separate initialization of the `RealTimeThread` to passing in the `BufferState` reference, and the current sample rate.

Since this makes a Big Function, add some comments delineating each step (& emphasizing the order-dependence), and add a `TRACEABLE_SCOPE`. While we're at it, validate that `jack_set_process_callback()` succeeded before continuing, and assert that any other unchecked maybe-invalid state is valid.