Add Meson project & options
!! THIS HAS NOT BEEN TESTED ON LINUX! IT WILL ALMOST CERTAINLY BREAK! I AM PUSHING THIS SO IT CAN BE TESTED ON LINUX !!

Meson can now be used to build the project on Windows, and technically on Linux, although this is completely untested and will assuredly break until all assuredly-hidden issues are fixed.

I'd like to add a couple of helper scripts & native config files to simplify the process, but for now, here's what to do:
- `pip install meson` and `pip install ninja`, ideally into a project-local virtual envrionment, if you've got one.
- Check out the `meson.options` file to see available build options.
- Pass these to `meson setup <all my options> build`, where each option is `-Doption=value`. E.g., `meson setup -Daudio_backend=jack -Dmidi_backend=alsa -Dtracy=enabled build`. This will generate a ready-to-go build script in a `build` subfolder.
- Run `meson compile -C build` to actually run the build.
` Run `meson install -C build` to move the generated mollytime extension to its final destination (in `./mollytime`).

`pybind11` is expected to be installed and available to Python. `boost-stacktrace` is ideally found from the system, but if not, you can put it in `third_party/boost-stacktrace-1.89.0` as a fallback, until there's a more complete model for third-party dependencies on Windows.