Add WASAPI backend (output only)
`#define AUDIO_WASAPI` to turn it on. The Meson build will handle this for you if `audio_backend` is set to `wasapi`.

WASAPI's scope is limited to interfacing with hardware endpoints, so the "input" and "aux" capabilities of the JACK backend -- which can pipe audio to and from any number of arbitrary "software" endpoints -- aren't straightforwardly applicable. We'll have to do something else to get that functionality on Windows, if desired.

There's no inherent "server" thread with WASAPI; we're responsible for execution control. This means that the `WasapiRealTimeThread` does genuinely need to run in a separate `std::thread` loop. Likewise, all of the COM objects we're handling are logically owned by the independent RealTimeThread that needs them, rather than the backend; the latter is thus limited to nothing more than main-thread ownership of the RealTimeThread and its BufferState.

The other major discrepancy is that WASAPI expects left/right stereo samples to be interleaved into consecutive "frames," rather than written independently. To prevent any higher-level changes, the backend takes care of it -- all samples are written into per-channel intermediate buffers, and then interleaved into the WASAPI output buffer on `EndFrame()`. This costs a little bit of extra memory (~2 KB), and presumably a little extra time (not measurably significant in Tracy).

Otherwise, implementation is straightforward:
- Initialize COM, and the minimal necessary WASAPI infrastructure. This includes getting the default `IMMDevice` (hardware endpoint), activating an `IAudioClient` (generic audio stream to the endpoint), for which we set up an `IAudioRenderClient` ("render", i.e. output, interface for the audio stream).
- Audio client initialization specifies "shared" mode -- i.e., not exclusive -- and uses an Event handle to signal when the buffer needs filling, so we don't have to manually calculate the timing ourselves.
- Once everything's ready to go, just loop the thread until shutdown. On shutdown, signal to the thread (via atomic bool) that we're done, and safely join before shutting down the audio stream.

Note: I'm using the WinRT `com_ptr` -- the latest in a long line of such -- to manage COM objects and their lifetimes. This is technically available on any WASAPI-capable Windows SDK, but on older SDK versions, it'll fail to compile in C++23 mode due to outdated experimental coroutine headers. Compiling with "Windows 11 SDK (10.0.22621.5040)" or newer is recommended. (Don't worry, it works just fine on Windows 10.) I've update the readme to reflect this.