Sketching Out New Select/Connect Screen
I want to make it so the select screen has shelves on the left and right
which have buttons for all the output and input ports for the current
selection.

Left hand side has the outputs from the first selection.  One of them
(first, by default) is always highlighted, and may be changed by clicking
on a different port.

Right hand side has the inputs on the second selection.  The ones that
currently have a connection to the selected lhs output port should all
be highlighted, and the connections can be created or broken by clicking
on the rhs ports.

There should also be arrows drawn between the ports representing all
current connections.  The ones corresponding to the current LHS selection
should be opaque, and the others should be semitransparent or a different
color.