Build: Don't include `boost_stacktrace` fallback if unavailable
I'm going to enable this by default in `debug` builds (where they're most valuable), but we don't currently guarantee availability of the library (unlike e.g. Tracy). I don't want to block compilation of debug builds on this, so allow the build to complete even if it can't be found.