Build: Use built-in  `extension_module`
I am dumb!! Meson's built-in Python module has first-class support for generating extension modules, which handles nearly all of the Python-specific wrangling the build script was previously doing:
- Making the Python `libs` path available on Windows.
- Getting the correct file suffix for the output library.
- Ensuring no `lib` prefix on the output library on non-Windows.
- Placing the output library in the correct location as an `install` step.

We don't have to do *any* of that. Look at the diff! Nearly all red! The only thing we have to do is specify `subdir` for the intended output location (otherwise Meson will put it in the user's system package location).