Removed Pointless Premature Optimization That Complicated Runtime Assumptions
This change replaces function MidiHandler::SwapMidiMessageQueue with the
simpler MidiHandler::PopMidiMessage function, and makes
MidiHandler::PendingMidiMessages a deque instead of a vector.

The bulk processing was completely unnecessary, as the throughput of the
MIDI protocol is substantially less than that of our realtime audio thread.

This change allows for the assumption that stuff like note tiles and a future
tile for CC events will always have at most one change between the last eval.

This removes the need for caching every possible control code per channel
regardless of if they are used in a patch, as well as eliminates the
possibility of note events cancelling out or otherwise dropping as a
consequence of batch processing.