MIDI Control Change Tile
I'm debating replacing this with a different implementation.

This implementation does the event filtering on the control change
tile, which means you need three tiles on the board to select unless
the input channel happens to be zero.

The shortcoming of this approach is the same as with the channel select
scheme all MIDI tiles currently use: you have to know what stuff is
sending in advance of writing your program, and there's no way to
inspect the MIDI stream directly without leaving mollytime.

One solution could be to have the filtering be explicit in the patch,
and provide a tile that just shows the value being passed into it.
Another approach could be to include a MIDI debug view somewhere.  I'm
undecided as to which is more ideal right now.