Gold Plated Sin Tile
This adds a software implementation of the sin tile as an
alternative to the regular sin tile which uses std::sin (which is
assumed to be a hardware implementation on systems with a FP ALU).

I can't actually hear the difference on my laptop though, so I'm
not merging this until I know of a compelling case for it.