Discard Truncated ELP for Worse Moon Approximation (WIP)
This is following the math from
https://aa.quae.nl/en/reken/hemelpositie.html#4, with the intention of
calculating the Moon's Alt/Az horizontal coordinates (or maybe just the
altitude angle part) to use to calculated the output of the tide tile.

I'm primarily interested in (AltitudeH / 90), where 1.0 is when the Moon
is at the zenith of the observer's sky, and 0.0 is when it is at the
horizon, etc.  I do not plan on exposing the azimuth, but I'll probably
still partially implement it should I need it later.

I'm also considering changing the name of this tile from "tide" to
"moon", since a proper tidal calculation also requires information about
the sun.  Or perhaps this one should be "moon height" and another
could be "moon phase".  If they would have the same inputs though, it
would just make sense to have one tile, and maybe use templates to
contextually enable or disable computations that are not relevant to
the connected inputs.