Very Basic Drum Machine Patch
I'm feeling out whether or not I want a toggle switch primitive, as
well as a periodic counting primitive and/or an equality test.  This
patch is meant to be a reference point for what life is like without
such things so I can get a sense of how beneficial these things are.

I'm leaning towards "yes" to a toggle switch since not seeing the
state of a flipflop is annoying when manipulating it directly, and
it would be nice to be able to save the active state of a toggle
switch, but I would not want to be able to do that for a flipflop.

I think the runtime cost of constructing a patch this way is
inconclusive, it fairly straight forward to make the other verbs out
of existing primitives, and the space reduction would probably not be
as much as I'd like.  I think it would therefor be better to focus my
energy on adding some kind of subroutine primitive instead of the
proposed new tile types mentioned above (besides toggle).

In practice, I do not think it is very useful to build a drum
sequencer directly into a patch like this, and I'd rather have a
proper sequencer mode built in instead in the future.