Deterministic Output Channel Assignment and Labeling
Output tiles are now labeled as being left or right outputs if they
are left or right outputs, and which output is left or right is now
consistent between patch loads.  Additionally, the assignment rules
are more logical for editing patches.

The assignment method is to sort the accepted output tiles by tile
ID.  If there are two or more, the lowest is assigned the left
channel, the second lowest is assigned the right channel, and the
remainder are labeled as being ignored.