Windows patch loading bug workaround
For some reason "load_path" would sometimes but not always be an empty
string when I'd try to open example "clock6.beep".  When I try to observe
the empty and non-empty case, the empty case never gets hit.  This change
seems to prevent the issue from showing up, but I am suspicious of this
solution.