Reworked ADSR to be more explicitly rate based.
In theory, this does not significantly change the behavior of ADSR unless Sustain < 1.
This corrects several edge cases where the ADSR could either get stuck above zero
during the release phase, and/or greatly prolong the release time beyond the expected
amount as Sustain approaches zero.

This changes bring ADSR's more behavior in line with or very close to my my original
intention of having ADR describe the time it takes to reach each phase's terminus.

I'm starting to think this model of the parameters is inferior to specifying ADR as
amplitude-per-second, but I need to do some more experimentation and research before
I decide switching to that is a good idea vs adding another adsr time.